+++ src/test/java/kitchenpos/application/ServiceTestHelper.java
@@ -6,7 +6,7 @@ import kitchenpos.application.menu.request.MenuCreateRequest;
 import kitchenpos.application.menu.request.MenuGroupCreateRequest;
 import kitchenpos.application.menu.request.MenuProductRequest;
 import kitchenpos.application.order.OrderService;
-import kitchenpos.application.tablegroup.TableService;
+import kitchenpos.application.table.TableService;
 import kitchenpos.application.order.request.OrderCreateRequest;
 import kitchenpos.application.order.request.OrderLineItemCreateRequest;
 import kitchenpos.application.order.request.OrderUpdateRequest;
@@ -14,17 +14,17 @@ import kitchenpos.application.order.request.TableCreateRequest;
 import kitchenpos.application.order.request.TableUpdateRequest;
 import kitchenpos.application.product.ProductService;
 import kitchenpos.application.product.request.ProductCreateRequest;
-import kitchenpos.application.tablegroup.TableGroupService;
-import kitchenpos.application.tablegroup.request.OrderTableIdRequest;
-import kitchenpos.application.tablegroup.request.TableGroupCreateRequest;
+import kitchenpos.application.table.TableGroupService;
+import kitchenpos.application.table.request.OrderTableIdRequest;
+import kitchenpos.application.table.request.TableGroupCreateRequest;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProductRepository;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.tablegroup.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.product.Product;
-import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.domain.table.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
