+++ src/test/java/kitchenpos/application/order/TableServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.application.order;
 
 import kitchenpos.application.ServiceTestHelper;
-import kitchenpos.domain.tablegroup.OrderTable;
-import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
