+++ src/test/java/kitchenpos/ui/table/TableGroupRestControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.tablegroup;
+package kitchenpos.ui.table;
 
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
